Read in data for a bin chunk

Read in data for a bin chunk

.pullBinChunks(
  binChunk,
  file,
  bin_offsets,
  binChunkSize,
  datasetPath,
  end1bin,
  start2bin,
  end2bin
)

Arguments

binChunk

The binChunk indeces to read

file

Path to .(m)cool file

bin_offsets

Read in bin1 offsets

binChunkSize

Size of bin chunk, for comparison against the end of the bin chunk

datasetPath

Dataset path, for specifying resolution in .mcool file

end1bin

Bin where end1 starts

start2bin

Bin for chr2 starts

end2bin

Bin for end2 starts

Value

An integer vector of pixel indices corresponding to interactions within the specified bin chunk, or NA if none are found.